// src/interfaces/enums/planillas.enums.ts
export enum TipoPeriodoPago {
  QUINCENAL = "quincenal",
  MENSUAL = "mensual",
  SEMANAL = "semanal",
}

export enum EstadoPlanilla {
  BORRADOR = "borrador",
  CONFIRMADA = "confirmada",
  PAGADA = "pagada",
  ANULADA = "anulada",
}

export enum MetodoPago {
  EFECTIVO = "efectivo",
  TRANSFERENCIA = "transferencia",
  CHEQUE = "cheque",
}